Found the bug that was destroying reconstruction
What is now `test/deltabot.test.coffee:61` after this commit, was the issue. I
had inadvertently switched the roles of the two test printers (ideal and actual)
when generating the test case. This means that there was no chance that the
test would ever pass.

Now, however, there's another issue: the reconstructed printer only takes into
account the z-discrepancy. I now need to take into account error in the x-y
plane. The reconstructed values are currently ~1mm off of actual values because
the reconstruction is not constrained by the x-y error.